+++ js/dataframe/shared/Table.svelte
@@ -170,6 +170,7 @@
 	let is_fullscreen = false;
 	let dragging = false;
 	let color_accent_copied: string;
+	let filtered_to_original_map: number[] = [];
 
 	onMount(() => {
 		const color = getComputedStyle(document.documentElement)
@@ -252,8 +253,18 @@
 
 	$: if ($df_state.current_search_query !== undefined) {
 		const cell_map = new Map();
+		filtered_to_original_map = [];
 
 		data.forEach((row, row_idx) => {
+			if (
+				row.some((cell) =>
+					String(cell?.value)
+						.toLowerCase()
+						.includes($df_state.current_search_query?.toLowerCase() || "")
+				)
+			) {
+				filtered_to_original_map.push(row_idx);
+			}
 			row.forEach((cell, col_idx) => {
 				cell_map.set(cell.id, {
 					value: cell.value,
@@ -275,6 +286,8 @@
 				};
 			})
 		);
+	} else {
+		filtered_to_original_map = [];
 	}
 
 	let previous_headers = _headers.map((h) => h.value);
@@ -891,7 +904,10 @@
 							<TableCell
 								bind:value={search_results[index][j].value}
 								display_value={get_cell_display_value(index, j)}
-								{index}
+								index={$df_state.current_search_query !== undefined &&
+								filtered_to_original_map[index] !== undefined
+									? filtered_to_original_map[index]
+									: index}
 								{j}
 								{actual_pinned_columns}
 								{get_cell_width}