+++ js/lite/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gradio/lite
 
+## 5.24.0
+
+### Dependency updates
+
+- @gradio/core@0.13.2
+- @gradio/atoms@0.15.0
+
 ## 5.23.1
 
 ### Dependency updates