+++ js/model3D/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @gradio/model3d
 
+## 0.14.8
+
+### Dependency updates
+
+- @gradio/statustracker@0.10.7
+- @gradio/atoms@0.15.0
+- @gradio/icons@0.11.0
+- @gradio/upload@0.16.0
+
 ## 0.14.7
 
 ### Features